+++ app/assets/js/scripts/landing.js
@@ -94,12 +94,14 @@ function hasRequiredBundleFiles(serverId) {
         // Iris / Sodium 번들이 추가되면 이전 캐시가 "완료"로 오판되지 않도록 필수 파일에 포함한다.
         { path: nodePath.join(commonModStore, 'mods', 'iris-neoforge-1.8.12+mc1.21.1.jar'), size: 2438548 },
         { path: nodePath.join(commonModStore, 'mods', 'pamhc2crops-NEOFORGE-1.21.1-1.0.0.jar'), size: 1047144 },
+        { path: nodePath.join(commonModStore, 'mods', 'pamhcpokevill-1.0.0.jar'), size: 1068841 },
         { path: nodePath.join(commonModStore, 'mods', 'pokemonquest-1.0.25.jar'), size: 2094771 },
         { path: nodePath.join(commonModStore, 'mods', 'pokevill-1.0.0.jar'), size: 89044 },
         { path: nodePath.join(commonModStore, 'mods', 'pokevillgacha-1.0.0 (3).jar'), size: 351133 },
         { path: nodePath.join(commonModStore, 'mods', 'sodium-neoforge-0.6.13+mc1.21.1.jar'), size: 1162994 },
         { path: nodePath.join(commonModStore, 'mods', 'voicechat-neoforge-1.21.1-2.6.12.jar'), size: 4890489 },
         { path: nodePath.join(commonModStore, 'mods', 'worldedit-mod-7.3.8.jar'), size: 6222854 },
+        { path: nodePath.join(instanceDir, 'resourcepacks', 'build.zip'), size: 4268569 },
         { path: nodePath.join(instanceDir, 'resourcepacks', 'pokevill.zip'), size: 30514651 },
         { path: nodePath.join(instanceDir, 'shaderpacks', 'ComplementaryUnbound_r5.5.1.zip'), size: 459175 }
     ]